receipt repository tests

+++ src/main/java/com/cookbook/cookbook/schemas/cookbook/Category.java
@@ -1,7 +1,6 @@
 package com.cookbook.cookbook.schemas.cookbook;
 
 import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -25,7 +24,7 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String name;
